Move FakeRemoteActivityStorate to new activitylistmodeltestutils
